refactor : Change tests dirs
