+++ src/App/Tests/index.handler.spec.ts
@@ -1,7 +1,7 @@
 import {InversifyExpressServer} from 'inversify-express-utils';
 import supertest from 'supertest';
 import {ICreateConnection} from '@digichanges/shared-experience';
-import initServer from './initServer';
+import initTestServer from '../../initTestServer';
 
 describe('Start Index Test', () =>
 {
@@ -11,7 +11,7 @@ describe('Start Index Test', () =>
 
     beforeAll(async(done) =>
     {
-        const configServer = await initServer();
+        const configServer = await initTestServer();
 
         server = configServer.server;
         request = configServer.request;
