+++ src/initTestServer.ts
@@ -1,39 +1,42 @@
 import 'reflect-metadata';
-import '../App/Presentation/Handlers/IndexHandler';
-import '../Auth/Presentation/Handlers/AuthHandler';
-import '../Item/Presentation/Handlers/ItemHandler';
-import '../User/Presentation/Handlers/UserHandler';
-import '../Role/Presentation/Handlers/RoleHandler';
-// import "../Presentation/Handlers/FileHandler";
-// import "../Presentation/Handlers/NotificationHandler";
-
-import {Locales} from '../app';
 
 import express from 'express';
 import bodyParser from 'body-parser';
 import compression from 'compression';
 import cors from 'cors';
 import helmet from 'helmet';
-import {ICreateConnection, ITokenRepository} from '@digichanges/shared-experience';
-
-import LoggerWinston from '../App/Presentation/Middlewares/LoggerWinston';
-import AuthenticationMiddleware from '../Auth/Presentation/Middlewares/AuthenticationMiddleware';
-import RefreshTokenMiddleware from '../App/Presentation/Middlewares/RefreshTokenMiddleware';
 
 import {InversifyExpressServer} from 'inversify-express-utils';
-import {ErrorHandler} from '../App/Presentation/Shared/ErrorHandler';
 import supertest from 'supertest';
-import DatabaseFactory from '../App/Infrastructure/Factories/DatabaseFactory';
-import SeedFactory from '../App/Infrastructure/Seeds/SeedFactory';
-import EventHandler from '../App/Infrastructure/Events/EventHandler';
-import RedirectRouteNotFoundMiddleware from '../App/Presentation/Middlewares/RedirectRouteNotFoundMiddleware';
-import {REPOSITORIES} from '../repositories';
-import TokenMongoRepository from '../Auth/Infrastructure/Repositories/TokenMongoRepository';
-import {validateEnv} from '../Config/validateEnv';
-import container from '../inversify.config';
-import ITokenDomain from '../App/InterfaceAdapters/ITokenDomain';
 
-const initServer = async(): Promise<any> =>
+import './App/Presentation/Handlers/IndexHandler';
+import './Auth/Presentation/Handlers/AuthHandler';
+import './Item/Presentation/Handlers/ItemHandler';
+import './User/Presentation/Handlers/UserHandler';
+import './Role/Presentation/Handlers/RoleHandler';
+// import "../Presentation/Handlers/FileHandler";
+// import "../Presentation/Handlers/NotificationHandler";
+
+import {Locales} from './app';
+
+import {ICreateConnection, ITokenRepository} from '@digichanges/shared-experience';
+
+import LoggerWinston from './App/Presentation/Middlewares/LoggerWinston';
+import AuthenticationMiddleware from './Auth/Presentation/Middlewares/AuthenticationMiddleware';
+import RefreshTokenMiddleware from './App/Presentation/Middlewares/RefreshTokenMiddleware';
+
+import {ErrorHandler} from './App/Presentation/Shared/ErrorHandler';
+import DatabaseFactory from './App/Infrastructure/Factories/DatabaseFactory';
+import SeedFactory from './App/Infrastructure/Seeds/SeedFactory';
+import EventHandler from './App/Infrastructure/Events/EventHandler';
+import RedirectRouteNotFoundMiddleware from './App/Presentation/Middlewares/RedirectRouteNotFoundMiddleware';
+import {REPOSITORIES} from './repositories';
+import TokenMongoRepository from './Auth/Infrastructure/Repositories/TokenMongoRepository';
+import {validateEnv} from './Config/validateEnv';
+import container from './inversify.config';
+import ITokenDomain from './App/InterfaceAdapters/ITokenDomain';
+
+const initTestServer = async(): Promise<any> =>
 {
     validateEnv();
 
@@ -93,4 +96,4 @@ const initServer = async(): Promise<any> =>
     return {request, dbConnection};
 };
 
-export default initServer;
+export default initTestServer;
